Remove compiler warnings

+++ simgear/math/SGMatrix.hxx
@@ -562,10 +562,10 @@ inline
 SGMatrixf
 toMatrixf(const SGMatrixd& m)
 {
-  return SGMatrixf(m(0,0), m(0,1), m(0,2), m(0,3),
-                   m(1,0), m(1,1), m(1,2), m(1,3),
-                   m(3,0), m(2,1), m(2,2), m(2,3),
-                   m(4,0), m(4,1), m(4,2), m(4,3));
+  return SGMatrixf((float)m(0,0), (float)m(0,1), (float)m(0,2), (float)m(0,3),
+                   (float)m(1,0), (float)m(1,1), (float)m(1,2), (float)m(1,3),
+                   (float)m(3,0), (float)m(2,1), (float)m(2,2), (float)m(2,3),
+                   (float)m(4,0), (float)m(4,1), (float)m(4,2), (float)m(4,3));
 }
 
 inline
